+++ src/Plugin.php
@@ -45,6 +45,7 @@ class Plugin extends BasePlugin
         if (class_exists(Assets::class) && !$app->getPlugins()->has('Assets')) {
             $plugin = new Assets();
             $plugin->bootstrap($app);
+            $plugin->disable('bootstrap');
             $app->addPlugin($plugin);
         }
     }